Reorganize modals
