+++ src/components/Modal/EpicDetailsModal/EpicDetailsModal.vue
@@ -1,0 +1,9 @@
+<template>
+	<div></div>
+</template>
+
+<script>
+export default {
+
+}
+</script>
